requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
google-genai==0.3.0
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
